backend/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.21.23
botocore==1.24.46
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
gunicorn==20.1.0
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
pyUFbr==0.1.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
traitlets==5.2.2.post1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.2.0
